requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
python-dotenv==0.10.3
requests==2.22.0
urllib3==1.25.7
HX711==1.1.2.3
numpy==1.17.4
smbus==1.1.post2
RPi.GPIO==0.7.0
